manage staff: - use common user repo/service. - new -> done class role-component.tsx. - done backend + front end manage staff.
