+++ src/main/java/com/resteam/smartway/repository/RoleRepository.java
@@ -4,7 +4,6 @@ import com.resteam.smartway.domain.Restaurant;
 import com.resteam.smartway.domain.Role;
 import com.resteam.smartway.security.multitenancy.annotation.DisableRestaurantFilter;
 import com.resteam.smartway.security.multitenancy.repository.BaseRepository;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
@@ -14,8 +13,4 @@ public interface RoleRepository extends BaseRepository<Role> {
     Role findByNameAndRestaurant(String name, Restaurant restaurant);
 
     Optional<Role> findOneByName(String name);
-
-    Role findByName(String name);
-
-    List<Role> findAllBy();
 }
