+++ src/main/java/com/resteam/smartway/repository/UserRepository.java
@@ -4,8 +4,14 @@ import com.resteam.smartway.domain.Restaurant;
 import com.resteam.smartway.domain.Role;
 import com.resteam.smartway.domain.User;
 import com.resteam.smartway.security.multitenancy.repository.BaseRepository;
+import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -22,4 +28,12 @@ public interface UserRepository extends BaseRepository<User> {
 
     @EntityGraph("user-with-authorities-entity-graph")
     Optional<User> findOneWithAuthoritiesByUsername(String username);
+
+    @Query(
+        "SELECT u FROM User u WHERE (:search IS NULL " +
+        "OR LOWER(u.fullName) LIKE CONCAT('%', :search, '%')" +
+        "OR LOWER(u.username) LIKE CONCAT('%', :search, '%')) " +
+        "AND (coalesce(:roleIdList) IS NULL OR u.role.id IN :roleIdList)"
+    )
+    Page<User> findWithFilterParams(@Param("search") String search, @Param("roleIdList") List<UUID> roleIdList, Pageable pageable);
 }
