+++ src/main/java/com/resteam/smartway/service/UserService.java
@@ -1,252 +1,33 @@
 package com.resteam.smartway.service;
 
-import com.resteam.smartway.config.Constants;
-import com.resteam.smartway.domain.Authority;
-import com.resteam.smartway.domain.Restaurant;
-import com.resteam.smartway.domain.Role;
 import com.resteam.smartway.domain.User;
-import com.resteam.smartway.repository.AuthorityRepository;
-import com.resteam.smartway.repository.RestaurantRepository;
-import com.resteam.smartway.repository.RoleRepository;
-import com.resteam.smartway.repository.UserRepository;
-import com.resteam.smartway.security.SecurityUtils;
-import com.resteam.smartway.security.multitenancy.context.RestaurantContext;
-import com.resteam.smartway.service.dto.AdminUserDTO;
+import com.resteam.smartway.service.dto.StaffDTO;
 import com.resteam.smartway.service.dto.TenantRegistrationDTO;
-import com.resteam.smartway.web.rest.errors.SubdomainAlreadyUsedException;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import tech.jhipster.security.RandomUtil;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
-/**
- * Service class for managing users.
- */
-@Service
-@Transactional
-public class UserService {
+public interface UserService {
+    Optional<User> completePasswordReset(String newPassword, String key);
 
-    private final Logger log = LoggerFactory.getLogger(UserService.class);
+    Optional<User> requestPasswordReset(String mail);
 
-    private final UserRepository userRepository;
+    void registerUser(TenantRegistrationDTO tenantRegistrationDTO);
 
-    private final PasswordEncoder passwordEncoder;
+    void changePassword(String currentClearTextPassword, String newPassword);
 
-    private final AuthorityRepository authorityRepository;
+    Optional<User> getUserWithAuthorities();
 
-    private final RestaurantRepository restaurantRepository;
+    List<String> getAuthorities();
 
-    private final RoleRepository roleRepository;
+    Page<StaffDTO> loadStaffsWithSearch(Pageable pageable, String searchText, List<String> roleIds);
 
-    public UserService(
-        UserRepository userRepository,
-        PasswordEncoder passwordEncoder,
-        AuthorityRepository authorityRepository,
-        RestaurantRepository restaurantRepository,
-        RoleRepository roleRepository
-    ) {
-        this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-        this.authorityRepository = authorityRepository;
-        this.restaurantRepository = restaurantRepository;
-        this.roleRepository = roleRepository;
-    }
+    StaffDTO createStaff(StaffDTO staffDTO);
 
-    public Optional<User> completePasswordReset(String newPassword, String key) {
-        log.debug("Reset user password for reset key {}", key);
-        return userRepository
-            .findOneByResetKey(key)
-            .filter(user -> user.getResetDate().isAfter(Instant.now().minus(1, ChronoUnit.DAYS)))
-            .map(user -> {
-                user.setPassword(passwordEncoder.encode(newPassword));
-                user.setResetKey(null);
-                user.setResetDate(null);
-                return user;
-            });
-    }
+    StaffDTO updateStaff(StaffDTO staffDTO);
 
-    public Optional<User> requestPasswordReset(String mail) {
-        return userRepository
-            .findOneByEmailIgnoreCase(mail)
-            .map(user -> {
-                user.setResetKey(RandomUtil.generateResetKey());
-                user.setResetDate(Instant.now());
-                return user;
-            });
-    }
+    void deleteStaff(List<String> ids);
 
-    public void registerUser(TenantRegistrationDTO tenantRegistrationDTO) {
-        restaurantRepository
-            .findOneById(tenantRegistrationDTO.getRestaurantId())
-            .ifPresent(r -> {
-                throw new SubdomainAlreadyUsedException();
-            });
-
-        Restaurant restaurant = new Restaurant(tenantRegistrationDTO.getRestaurantId());
-        Restaurant savedRestaurant = restaurantRepository.save(restaurant);
-        RestaurantContext.setCurrentRestaurant(savedRestaurant);
-
-        Role role = roleRepository.findByNameAndRestaurant("ADMIN", new Restaurant("system@"));
-
-        User newUser = new User();
-        String encryptedPassword = passwordEncoder.encode(tenantRegistrationDTO.getPassword());
-        newUser.setUsername(tenantRegistrationDTO.getUsername().toLowerCase());
-        newUser.setPassword(encryptedPassword);
-        newUser.setFullName(tenantRegistrationDTO.getFullName());
-        newUser.setEmail(tenantRegistrationDTO.getEmail().toLowerCase());
-        newUser.setPhone(tenantRegistrationDTO.getPhone());
-        newUser.setLangKey(tenantRegistrationDTO.getLangKey());
-        newUser.setRole(role);
-        userRepository.save(newUser);
-
-        log.debug("Created Information for User: {}", newUser);
-    }
-
-    private boolean removeNonActivatedUser(User existingUser) {
-        userRepository.delete(existingUser);
-        userRepository.flush();
-        return true;
-    }
-
-    public User createUser(AdminUserDTO userDTO) {
-        User user = new User();
-        user.setUsername(userDTO.getUsername().toLowerCase());
-        user.setFullName(userDTO.getFullName());
-        if (userDTO.getEmail() != null) {
-            user.setEmail(userDTO.getEmail().toLowerCase());
-        }
-        if (userDTO.getLangKey() == null) {
-            user.setLangKey(Constants.DEFAULT_LANGUAGE); // default language
-        } else {
-            user.setLangKey(userDTO.getLangKey());
-        }
-        String encryptedPassword = passwordEncoder.encode(RandomUtil.generatePassword());
-        user.setPassword(encryptedPassword);
-        user.setResetKey(RandomUtil.generateResetKey());
-        user.setResetDate(Instant.now());
-        if (userDTO.getAuthorities() != null) {
-            Set<Authority> authorities = userDTO
-                .getAuthorities()
-                .stream()
-                .map(authorityRepository::findById)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(Collectors.toSet());
-            user.setRole(new Role(null, "", authorities));
-        }
-        userRepository.save(user);
-        log.debug("Created Information for User: {}", user);
-        return user;
-    }
-
-    /**
-     * Update all information for a specific user, and return the modified user.
-     *
-     * @param userDTO user to update.
-     * @return updated user.
-     */
-    public Optional<AdminUserDTO> updateUser(AdminUserDTO userDTO) {
-        return Optional
-            .of(userRepository.findById(userDTO.getId()))
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .map(user -> {
-                user.setUsername(userDTO.getUsername().toLowerCase());
-                user.setFullName(userDTO.getFullName());
-                if (userDTO.getEmail() != null) {
-                    user.setEmail(userDTO.getEmail().toLowerCase());
-                }
-                user.setLangKey(userDTO.getLangKey());
-                Collection<Authority> managedAuthorities = user.getRole().getAuthorities();
-                managedAuthorities.clear();
-                userDTO
-                    .getAuthorities()
-                    .stream()
-                    .map(authorityRepository::findById)
-                    .filter(Optional::isPresent)
-                    .map(Optional::get)
-                    .forEach(managedAuthorities::add);
-                log.debug("Changed Information for User: {}", user);
-                return user;
-            })
-            .map(AdminUserDTO::new);
-    }
-
-    public void deleteUser(String login) {
-        userRepository
-            .findOneByUsername(login)
-            .ifPresent(user -> {
-                userRepository.delete(user);
-                log.debug("Deleted User: {}", user);
-            });
-    }
-
-    /**
-     * Update basic information (first name, last name, email, language) for the current user.
-     *
-     * @param firstName first name of user.
-     * @param email     email id of user.
-     * @param langKey   language key.
-     */
-    public void updateUser(String firstName, String email, String langKey) {
-        SecurityUtils
-            .getCurrentUsername()
-            .flatMap(userRepository::findOneByUsername)
-            .ifPresent(user -> {
-                user.setFullName(firstName);
-                if (email != null) {
-                    user.setEmail(email.toLowerCase());
-                }
-                user.setLangKey(langKey);
-                log.debug("Changed Information for User: {}", user);
-            });
-    }
-
-    @Transactional
-    public void changePassword(String currentClearTextPassword, String newPassword) {
-        SecurityUtils
-            .getCurrentUsername()
-            .flatMap(userRepository::findOneByUsername)
-            .ifPresent(user -> {
-                String currentEncryptedPassword = user.getPassword();
-                if (!passwordEncoder.matches(currentClearTextPassword, currentEncryptedPassword)) {
-                    throw new InvalidPasswordException();
-                }
-                String encryptedPassword = passwordEncoder.encode(newPassword);
-                user.setPassword(encryptedPassword);
-                log.debug("Changed password for User: {}", user);
-            });
-    }
-
-    //        @Transactional(readOnly = true)
-    //        public Page<UserDTO> getAllPublicUsers(Pageable pageable) {
-    //            return restaurantRepository.findAllByIdNotNullAndActivatedIsTrue(pageable).map(UserDTO::new);
-    //        }
-
-    @Transactional(readOnly = true)
-    public Optional<User> getUserWithAuthorities() {
-        String username = SecurityUtils.getCurrentUsername().orElseThrow(() -> new UsernameNotFoundException(("Username must be provide")));
-
-        return userRepository.findOneWithAuthoritiesByUsername(username);
-    }
-
-    /**
-     * Gets a list of all the authorities.
-     *
-     * @return a list of all the authorities.
-     */
-    @Transactional(readOnly = true)
-    public List<String> getAuthorities() {
-        return authorityRepository.findAll().stream().map(Authority::getName).collect(Collectors.toList());
-    }
+    void updateUser(String fullName, String email, String langKey);
 }
