+++ src/main/java/com/resteam/smartway/service/dto/StaffDTO.java
@@ -1,27 +1,24 @@
 package com.resteam.smartway.service.dto;
 
+import static com.resteam.smartway.service.dto.TenantRegistrationDTO.PASSWORD_MAX_LENGTH;
+import static com.resteam.smartway.service.dto.TenantRegistrationDTO.PASSWORD_MIN_LENGTH;
+
 import com.resteam.smartway.config.Constants;
-import com.resteam.smartway.domain.Role;
-import java.time.Instant;
-import java.util.Set;
 import java.util.UUID;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class StaffDTO {
-
-    private static final long serialVersionUID = 1L;
-
-    public static final int PASSWORD_MIN_LENGTH = 4;
-
-    public static final int PASSWORD_MAX_LENGTH = 100;
 
     private UUID id;
