+++ src/main/java/com/resteam/smartway/web/rest/AccountResource.java
@@ -8,11 +8,11 @@ import com.resteam.smartway.service.UserService;
 import com.resteam.smartway.service.dto.AdminUserDTO;
 import com.resteam.smartway.service.dto.PasswordChangeDTO;
 import com.resteam.smartway.service.dto.TenantRegistrationDTO;
-import com.resteam.smartway.web.rest.errors.*;
+import com.resteam.smartway.web.rest.errors.EmailAlreadyUsedException;
+import com.resteam.smartway.web.rest.errors.InvalidPasswordException;
 import com.resteam.smartway.web.rest.vm.KeyAndPasswordVM;
 import com.resteam.smartway.web.rest.vm.ManagedUserVM;
-import java.util.*;
-import javax.servlet.http.HttpServletRequest;
+import java.util.Optional;
 import javax.validation.Valid;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
