--- src/main/webapp/app/pages/user/management/role-staff/role.reducer.ts
@@ -1,122 +1,0 @@
-import axios from 'axios';
-import { createAsyncThunk, isFulfilled, isPending, isRejected } from '@reduxjs/toolkit';
-
-import { cleanEntity } from 'app/shared/util/entity-utils';
-import { IQueryParams, createEntitySlice, EntityState, serializeAxiosError } from 'app/shared/reducers/reducer.utils';
-import { defaultValue, IRole } from 'app/shared/model/role';
-
-const initialState: EntityState<IRole> = {
-  loading: false,
-  errorMessage: null,
-  entities: [],
-  entity: defaultValue,
-  updating: false,
-  updateSuccess: false,
-};
-
-const apiUrl = 'api/roles';
-
-// Actions
-
-export const getEntities = createAsyncThunk('role/fetch_entity_list', async ({ sort, page, size }: IQueryParams) => {
-  return axios.get<IRole[]>(apiUrl);
-});
-
-export const getEntity = createAsyncThunk(
-  'role/fetch_entity',
-  async (id: any) => {
-    const requestUrl = `${apiUrl}/${id}`;
-    return axios.get<IRole>(requestUrl);
-  },
-  { serializeError: serializeAxiosError }
-);
-
-export const createEntity = createAsyncThunk(
-  'role/create_entity',
-  async (entity: IRole, thunkAPI) => {
-    const result = await axios.post<IRole>(apiUrl, cleanEntity(entity));
-    thunkAPI.dispatch(getEntities({}));
-    return result;
-  },
-  { serializeError: serializeAxiosError }
-);
-
-export const updateEntity = createAsyncThunk(
-  'role/update_entity',
-  async (entity: IRole, thunkAPI) => {
-    const result = await axios.put<IRole>(`${apiUrl}/${entity.id}`, cleanEntity(entity));
-    thunkAPI.dispatch(getEntities({}));
-    return result;
-  },
-  { serializeError: serializeAxiosError }
-);
-
-export const partialUpdateEntity = createAsyncThunk(
-  'role/partial_update_entity',
-  async (entity: IRole, thunkAPI) => {
-    const result = await axios.patch<IRole>(`${apiUrl}/${entity.id}`, cleanEntity(entity));
-    thunkAPI.dispatch(getEntities({}));
-    return result;
-  },
-  { serializeError: serializeAxiosError }
-);
-
-export const deleteEntity = createAsyncThunk(
-  'role/delete_entity',
-  async (id: any, thunkAPI) => {
-    const requestUrl = `${apiUrl}/${id}`;
-    const result = await axios.delete<IRole>(requestUrl);
-    thunkAPI.dispatch(getEntities({}));
-    return result;
-  },
-  { serializeError: serializeAxiosError }
-);
-
-// slice
-
-export const RoleSlice = createEntitySlice({
-  name: 'role',
-  initialState,
-  extraReducers(builder) {
-    builder
-      .addCase(getEntity.fulfilled, (state, action) => {
-        state.loading = false;
-        state.entity = action.payload.data;
-      })
-      .addCase(deleteEntity.fulfilled, state => {
-        state.updating = false;
-        state.updateSuccess = true;
-        state.entity = {};
-      })
-      .addMatcher(isFulfilled(getEntities), (state, action) => {
-        const { data } = action.payload;
-
-        return {
-          ...state,
-          loading: false,
-          entities: data,
-        };
-      })
-      .addMatcher(isFulfilled(createEntity, updateEntity, partialUpdateEntity), (state, action) => {
-        state.updating = false;
-        state.loading = false;
-        state.updateSuccess = true;
-        state.entity = {};
-      })
-      .addMatcher(isPending(getEntities, getEntity), state => {
-        state.errorMessage = null;
-        state.updateSuccess = false;
-        state.loading = true;
-      })
-      .addMatcher(isPending(createEntity, updateEntity, partialUpdateEntity, deleteEntity), state => {
-        state.errorMessage = null;
-        state.updateSuccess = false;
-        state.updating = true;
-      });
-  },
-});
-
-export const { reset } = RoleSlice.actions;
-
-// Reducer
-export default RoleSlice.reducer;
