+++ src/main/webapp/app/pages/user/management/staff/staff.reducer.ts
@@ -6,9 +6,7 @@ import getStore from 'app/config/store';
 import { DEFAULT_PAGEABLE } from 'app/app.constant';
 import { cleanEntity, getListValuesInParam } from 'app/shared/util/entity-utils';
 import React from 'react';
-import { IMenuItem } from 'app/shared/model/menu-item.model';
 import { updateIsActiveEntity } from 'app/pages/user/management/menu-item/menu-item.reducer';
-import { IRole } from 'app/shared/model/role.model';
 
 const initialState: EntityState<IStaff> = {
   loading: false,
@@ -27,9 +25,9 @@ export const setPageable = createAsyncThunk('staff/set_pageable', (pageable: IQu
 });
 
 export const getEntities = createAsyncThunk('/staff/fetch_entity_list', async () => {
-  const { sort, page, size, restaurant, search } = getStore().getState().staff.pageable;
-  const requestUrl = `${apiUrl}?page=${page}&size=${size}&sort=${sort}&search=${search ? search : ''}&restaurantId=${getListValuesInParam(
-    restaurant
+  const { sort, page, size, role, search } = getStore().getState().staff.pageable;
+  const requestUrl = `${apiUrl}?page=${page}&size=${size}&sort=${sort}&search=${search ? search : ''}&roleIds=${getListValuesInParam(
+    role
   )}`;
   return axios.get<IStaff[]>(requestUrl);
 });
